nflMatchupPredictor/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
passlib==1.7.4
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
zipp==3.8.0
